fix : 증복 컴포넌트 분리
