+++ src/1_Page/Championship/ui/DashBoard/ui/TeamAndMatchContainerProps/ui/MatchListContainer/ui/MatchLineup/ui/TeamStatCard/constant/teamStatKeys.ts
@@ -1,0 +1,16 @@
+// 미리 정의한 팀 스탯 항목 목록 (TeamStats 타입 기준)
+export const teamStatKeys: TeamStatKey[] = [
+  { label: "Our Score", key: "match_team_stats_our_score" },
+  { label: "Other Score", key: "match_team_stats_other_score" },
+  { label: "Possession", key: "match_team_stats_possession" },
+  { label: "Total Shot", key: "match_team_stats_total_shot" },
+  { label: "Expected Goal", key: "match_team_stats_expected_goal" },
+  { label: "Total Pass", key: "match_team_stats_total_pass" },
+  { label: "Total Tackle", key: "match_team_stats_total_tackle" },
+  { label: "Successful Tackle", key: "match_team_stats_success_tackle" },
+  { label: "Saved", key: "match_team_stats_saved" },
+  { label: "Corner Kick", key: "match_team_stats_cornerkick" },
+  { label: "Free Kick", key: "match_team_stats_freekick" },
+  { label: "Penalty Kick", key: "match_team_stats_penaltykick" },
+  { label: "MOM Player", key: "mom_player_nickname" },
+];
