+++ src/1_Page/Championship/ui/DashBoard/ui/TeamAndMatchContainerProps/ui/MatchListContainer/ui/MatchLineup/ui/TeamStatCard/type.d.ts
@@ -1,0 +1,6 @@
+type TeamStatCardProps = {
+  teamName: string;
+  stats: TeamStats;
+  backgroundColorClass: string;
+  textColorClass: string;
+};
